requirements.txt
# requirements.txt

# Основные зависимости
pydub>=0.25.1
PyTelegramBotAPI>=4.14.0

# ASR провайдеры
# OpenAI Whisper (по умолчанию)
openai-whisper>=20231117

# NVIDIA NeMo (опционально)
# Раскомментируйте для использования NeMo
# nemo_toolkit[asr]>=1.20.0

# Дополнительные зависимости для обработки аудио
torch>=2.0.0
torchaudio>=2.0.0

# Для улучшенной производительности
ffmpeg-python>=0.2.0

# Для обработки путей
pathlib2>=2.3.7